Use generic task type references for prompt examples instead of kwargs. Drop Python 3.6 support.
